+++ src/components/layout/AuthenticatedLayout.tsx
